+++ bin/generate_code.sh
@@ -1,5 +1,0 @@
-#!/usr/bin/env sh
-
-flutter pub run build_runner build --delete-conflicting-outputs
-
-