+++ tool/generate_code.sh
@@ -1,0 +1,5 @@
+#!/usr/bin/env sh
+
+dart pub run build_runner build --delete-conflicting-outputs
+
+