+++ tool/publish.sh
@@ -1,0 +1,4 @@
+#!/usr/bin/env sh
+
+flutter pub publish
+