+++ CHANGELOG.md
@@ -1,3 +1,21 @@
-## 1.0.0
+## 0.1.0
 
-- Initial version.
+- Wikipedia Search, Summary with lang parameter
+
+## 0.2.0
+
+- Wikidata GetEntities
+
+## 0.3.0
+
+- Wikipedia Get Page Props
+
+# 0.3.1
+
+- Fix Readme
+
+# 0.3.2
+
+- Get entities improved
+- Api documentation
+